+++ CHANGELOG.rst
@@ -19,7 +19,7 @@ latest
 
 - Maintenance
 
-  - Testing: added Python 3.13.
+  - Testing: added Python 3.13; moved Mac/Win to 3.13.
 
 
 v2.4.0 El. current density